refactor(alignment): split alignment.rs god-file into src/alignment/ modules
Split the 1,893-line alignment.rs into args, io, header, model, stream,
fetch, and util submodules along natural seams, with tests moved next
to their units. Pure code movement plus visibility adjustments only;
no behaviour change (verified by full test suite + reproducibility
tests passing unchanged).
