+++ src/alignment/io.rs
@@ -1,0 +1,153 @@
+use std::fs::File;
+use std::io::{self, Write};
+use std::path::Path;
+
+use anyhow::{Context, Result};
+use log::info;
+use rand::prelude::*;
+use rand::random;
+
+use noodles::sam::Header;
+use noodles_util::alignment::{self, io::Format};
+
+use super::args::Alignment;
+
+// a generic writer that can handle File or Stdout
+pub(super) type AlignmentWriter = alignment::io::Writer<Box<dyn Write>>;
+
+impl Alignment {
+    pub(super) fn setup_resources(
+        &self,
+        input_header: &Header,
+    ) -> Result<(rand_pcg::Pcg64, AlignmentWriter)> {
+        // set up random number generator
+        let rng = match self.seed {
+            Some(s) => rand_pcg::Pcg64::seed_from_u64(s),
+            None => {
+                let seed = random();
+                info!("Using seed: {seed}");
+                rand_pcg::Pcg64::seed_from_u64(seed)
+            }
+        };
+
+        let mut header = input_header.clone();
+
+        // add rasusa program command line to header
+        let (pg_id, pg_map) = self.program_entry(&header);
+
+        // set up the header and writer
+        header.programs_mut().as_mut().insert(pg_id.into(), pg_map);
+
+        let input_fmt = match infer_format_from_path(&self.aln) {
+            Some(fmt) => fmt,
+            None => {
+                return Err(anyhow::anyhow!(
+                    "Output file format not recognized. Please use .sam, .bam, or .cram extensions"
+                ));
+            }
+        };
+
+        let output_fmt = match &self.output_format {
+            Some(fmt) => *fmt,
+            None => match &self.output {
+                None => input_fmt,
+                Some(path) => match infer_format_from_path(path) {
+                    Some(fmt) => fmt,
+                    None => {
+                        return Err(anyhow::anyhow!(
+                            "Output file format not recognized. Please use .sam, .bam, or .cram extensions"
+                        ));
+                    }
+                },
+            },
+        };
+
+        // use Box<dyn Write> to make File and Stdout compatible
+        let sink: Box<dyn Write> = match &self.output {
+            Some(path) => {
+                let path = path.as_path();
+                info!("Writing subsampled alignment to: {:?}", path);
+                let file = File::create(path).context("Failed to create output alignment file")?;
+                Box::new(file)
+            }
+            None => {
+                info!("Writing subsampled alignment to stdout");
+                Box::new(io::stdout().lock())
+            }
+        };
+
+        let mut writer = alignment::io::writer::Builder::default()
+            .set_format(output_fmt)
+            .build_from_writer(sink)?;
+
+        writer.write_header(&header)?;
+
+        Ok((rng, writer))
+    }
+}
+
+pub fn infer_format_from_path(path: &Path) -> Option<Format> {
+    match path.extension().and_then(|ext| ext.to_str()) {
+        Some("sam") => Some(Format::Sam),
+        Some("bam") => Some(Format::Bam),
+        Some("cram") => Some(Format::Cram),
+        _ => None,
+    }
+}
+
+// a function which infers a format from a single character (case insensitive) this will be used in the CLI, so should return a Result
+pub fn infer_format_from_char(c: &str) -> Result<Format, String> {
+    match c.to_ascii_lowercase().as_str() {
+        "s" | "sam" => Ok(Format::Sam),
+        "b" | "bam" => Ok(Format::Bam),
+        "c" | "cram" => Ok(Format::Cram),
+        _ => Err(String::from("Invalid output format. Please use 's', 'b', or 'c' to specify SAM, BAM, or CRAM format, respectively")),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_infer_format() {
+        let fmt = infer_format_from_path(Path::new("file.sam")).unwrap();
+        assert!(matches!(fmt, Format::Sam));
+
+        let fmt = infer_format_from_path(Path::new("file.bam")).unwrap();
+        assert!(matches!(fmt, Format::Bam));
+
+        let fmt = infer_format_from_path(Path::new("file.cram")).unwrap();
+        assert!(matches!(fmt, Format::Cram));
+
+        let fmt = infer_format_from_path(Path::new("file.txt"));
+        assert!(fmt.is_none());
+    }
+
+    #[test]
+    fn test_infer_format_from_char() {
+        let fmt = infer_format_from_char("s").unwrap();
+        assert!(matches!(fmt, Format::Sam));
+
+        let fmt = infer_format_from_char("b").unwrap();
+        assert!(matches!(fmt, Format::Bam));
+
+        let fmt = infer_format_from_char("c").unwrap();
+        assert!(matches!(fmt, Format::Cram));
+
+        let fmt = infer_format_from_char("x");
+        assert!(fmt.is_err());
+
+        let fmt = infer_format_from_char("sam").unwrap();
+        assert!(matches!(fmt, Format::Sam));
+
+        let fmt = infer_format_from_char("B").unwrap();
+        assert!(matches!(fmt, Format::Bam));
+
+        let fmt = infer_format_from_char("CRAM").unwrap();
+        assert!(matches!(fmt, Format::Cram));
+
+        let fmt = infer_format_from_char("x");
+        assert!(fmt.is_err());
+    }
+}
